Separate functionality into controllers directory
